pyproject.toml
[tool.poetry]
name = "scout-annotation"
version = "0.7.3"
description = "Snakemake pipeline with utilities for annotating VCFs for loading into Scout"
authors = ["Niklas Mähler <niklas.mahler@gmail.com>"]
license = "GPLv3"
readme = "README.md"
packages = [{include = "scout_annotation"}]
include = [
    "scout_annotation/data",
    "scout_annotation/default_config"
]

[tool.poetry.scripts]
scout-annotation = "scout_annotation.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
snakemake = "^7.13"
click = "^8.1.7"
cyvcf2 = "^0.30"
pysam = "^0.21"
pandas = "^1.4"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
